docker-clyde-network
Reads the docker yaml files in the base/ directory and creates the clyde
network and services that will ultimately execute on the network. The network
access is setup to be via a Command Line Interface at present that executes a
'bash' shell on the peer0.clyde.com - the first peer in the Clyde Hotel
oragnisation.